byte order independent code, transportRevision is int8

+++ pvAccessApp/remote/blockingClientTCPTransport.cpp
@@ -34,7 +34,7 @@ namespace epics {
         BlockingClientTCPTransport::BlockingClientTCPTransport(
                 Context::shared_pointer const & context, SOCKET channel,
                 auto_ptr<ResponseHandler>& responseHandler, int receiveBufferSize,
-                TransportClient::shared_pointer client, short remoteTransportRevision,
+                TransportClient::shared_pointer client, int8 remoteTransportRevision,
                 float beaconInterval, int16 priority) :
             BlockingTCPTransport(context, channel, responseHandler, receiveBufferSize, priority),
                     _introspectionRegistry(false),
