+++ pvAccessApp/remote/blockingUDPConnector.cpp
@@ -23,7 +23,7 @@ namespace epics {
 
     Transport::shared_pointer BlockingUDPConnector::connect(TransportClient::shared_pointer const & client,
                 auto_ptr<ResponseHandler>& responseHandler, osiSockAddr& bindAddress,
-                short transportRevision, int16 priority) {
+                int8 transportRevision, int16 priority) {
                     
             LOG(logLevelDebug, "Creating datagram socket to: %s",
                     inetAddressToString(bindAddress).c_str());
